Autolayout class decoupled into several little classes for better readability
