+++ Sources/LayoutKit/LayoutAnchor.swift
@@ -1,0 +1,16 @@
+//
+//  File.swift
+//  
+//
+//  Created by LEMIN DAHOVICH on 20.04.2023.
+//
+
+import UIKit
+
+public protocol LayoutAnchor: AnyObject {
+    func constraint(equalTo anchor: Self, constant: CGFloat) -> NSLayoutConstraint
+    func constraint(greaterThanOrEqualTo anchor: Self, constant: CGFloat) -> NSLayoutConstraint
+    func constraint(lessThanOrEqualTo anchor: Self, constant: CGFloat) -> NSLayoutConstraint
+}
+
+extension NSLayoutAnchor: LayoutAnchor {}
